requirements.txt
streamlit
langgraph

pandas
python-dotenv
pydantic
langchain
langchain-community
langchain-google-genai

-e .